Move related method to the web_ui_gherkin
