+++ ai_powered_qa/components/agent_store.py
@@ -108,35 +108,3 @@ class AgentStore:
         with open(file_path, "w") as file:
             file.write(interaction.model_dump_json(indent=4))
 
-    def load_gherkin_memory(self, agent: Agent):
-        file_name = "gherkin_memory.json"
-        history_directory = os.path.join(
-            self._directory, agent.agent_name, agent.history_name
-        )
-        file_path = os.path.join(history_directory, file_name)
-        if not os.path.exists(file_path):
-            return "No data"
-
-        with open(file_path, "r") as file:
-            return json.load(file)
-
-    def update_gherkin_memory(self, agent: Agent, property_name, new_value):
-        file_name = "gherkin_memory.json"
-        history_directory = os.path.join(
-            self._directory, agent.agent_name, agent.history_name
-        )
-        file_path = os.path.join(history_directory, file_name)
-
-        try:
-            with open(file_path, "r") as file:
-                data = json.load(file)
-        except FileNotFoundError:
-            data = {"gherkin_steps_history": [], "html_content": "`", "main_task": ""}
-            os.makedirs(history_directory, exist_ok=True)
-
-        if property_name in data and isinstance(data[property_name], list):
-            data[property_name].extend(new_value)
-        else:
-            data[property_name] = new_value
-        with open(file_path, "w") as file:
-            json.dump(data, file, indent=2)
